requirements.txt
requests==2.32.4
python-dotenv~=1.1.1
langchain
langchain-openai
langchain-community
sentence-transformers
faiss-cpu
pandas==2.2.2
langchain-huggingface
streamlit
